ref(base/participant): cleanup on CONFERENCE_WILL_LEAVE
Because the process of leaving a conference is asynchronous and there's
no guarantee on when and if the PARTICIPANT_LEFT events will come,
the participants have to be cleaned up as soon as we know that we're
going to leave the conference.

This commit also bring the local participant ID update earlier in the
events chain to change it on CONFERENCE_WILL_LEAVE. Otherwise weird
things happen.For example in case of a failed conference rejoin the app
thinks that our ghost is a local participant.

Another thing brought by this commit is a reference to the parent
conference added to each participant state stored in redux.